#2612 use more meaningful names for character length parameters

+++ src/psyclone/tests/psyir/backend/fortran_unsupported_declns_test.py
@@ -263,8 +263,10 @@ def test_generating_unsupportedtype_routine_imports(
               use a_mod, only: unsupported_type_symbol
               contains
               subroutine test()
-                  integer :: a
-                  a = INT(REAL(unsupported_type_symbol()))
+                  real :: a
+                  ! The function call returns a complex number and we want
+                  ! to store a simple scalar.
+                  a = REAL(unsupported_type_symbol())
               end subroutine test
           end module test_mod
       ''')
